Add MacOS Build Instructions
Here we add detailed build instructions to make it possible to build the ash compiler on MacOS from source using `ghcup` and `cabal`. There are a number of small changes, but they are necessary for a proper build. We document this in a new section of the MacOS README for Ash.