Add Basic Logo and License
This patch adds a simple logo and license for Ash lang. We just rename it to Ash and add a simple logo. Finally, we remove the outmoded github items, which we will replace with better ones (e.g. unit tests). Lastly, we add CODEOWNERs.

Note: we'll need to update the link when this is merged to the main branch, but I wanted to make sure it looks visually okay.

Note: we retain the Elm license, as required by BSD-3, but also attach our own (identical BSD-3 license). We may be able to unify them, but I'd need to specifically research that.